refactor(test): consolidate saturation tests into single file
Merged saturation_progression_test.go into saturation_test.go to keep all
saturation-related tests in one place. This improves test organization and
reduces file count without changing any test logic.

- Moved TestSaturationProgression_RealWorkloads and generateTestWorkload()
  into saturation_test.go
- Deleted sim/workload/saturation_progression_test.go
- All tests still passing
