--- sim/workload/saturation_progression_test.go
@@ -1,174 +1,0 @@
-package workload
-
-import (
-	"fmt"
-	"testing"
-	"time"
-
-	sim "github.com/inference-sim/inference-sim/sim"
-)
-
-// TestSaturationProgression_RealWorkloads demonstrates the classification progression
-// from UNSATURATED → TRANSIENT_BACKLOG → PERSISTENTLY_SATURATED using actual simulator
-// runs at increasing request rates.
-//
-// Key insight: PERSISTENTLY_SATURATED requires that observation ends BEFORE the workload
-// drains. Finite closed-loop workloads will always eventually drain, causing negative
-// slopes if observed through completion. Use --horizon to truncate observation during
-// the load phase.
-//
-// This test documents the expected behavior by running real simulations, not synthetic data.
-func TestSaturationProgression_RealWorkloads(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping real workload test in short mode")
-	}
-
-	cfg := NewBacklogDriftConfig(
-		10*time.Second, // window size
-		4,              // min windows  
-		2.0,            // peak ratio threshold
-		0.2,            // peak ratio band
-		0.95,           // confidence level
-	)
-	analyzer := NewBacklogDriftAnalyzer(cfg)
-
-	// Test cases demonstrate saturation progression
-	// Rate increases → classification severity increases
-	tests := []struct {
-		name              string
-		rate              float64
-		numRequests       int
-		horizonUs         int64
-		expectedClass     string
-		expectPositiveCI  bool // CI lower bound > 0
-	}{
-		{
-			name:             "Low rate - comfortable capacity",
-			rate:             5.0,
-			numRequests:      500,
-			horizonUs:        0, // Run to completion
-			expectedClass:    "UNSATURATED",
-			expectPositiveCI: false,
-		},
-		{
-			name:             "Medium rate - approaching saturation",
-			rate:             100.0,
-			numRequests:      10000,
-			horizonUs:        60_000_000, // Truncate before drain
-			expectedClass:    "TRANSIENT_BACKLOG",
-			expectPositiveCI: false, // Slope near zero, high peak/mean
-		},
-		{
-			name:             "High rate - persistent saturation",
-			rate:             500.0,
-			numRequests:      100000,
-			horizonUs:        120_000_000, // Truncate before drain
-			expectedClass:    "PERSISTENTLY_SATURATED",
-			expectPositiveCI: true, // CI excludes zero
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Generate workload
-			requests := generateTestWorkload(tt.rate, tt.numRequests, tt.horizonUs)
-
-			// Compute simEndUs
-			simEndUs := tt.horizonUs
-			if simEndUs == 0 {
-				// Use actual completion times
-				for _, req := range requests {
-					if req.TTFTSet {
-						completionUs := req.ArrivalTime + req.FirstTokenTime
-						for _, itl := range req.ITL {
-							completionUs += itl
-						}
-						if completionUs > simEndUs {
-							simEndUs = completionUs
-						}
-					}
-				}
-			}
-
-			// Analyze
-			report := analyzer.Analyze(requests, simEndUs)
-
-			// Verify classification
-			if report.Classification != tt.expectedClass {
-				t.Errorf("Rate %.1f req/s:\n  Expected: %s\n  Got:      %s\n  Note:     %s",
-					tt.rate, tt.expectedClass, report.Classification, report.Note)
-			}
-
-			// Verify slope CI behavior
-			data := report.AlgorithmData.(BacklogDriftData)
-			ciExcludesZero := data.SlopeLower > 0
-			if tt.expectPositiveCI && !ciExcludesZero {
-				t.Errorf("Rate %.1f req/s: Expected CI to exclude zero, got CI=[%.6e, %.6e]",
-					tt.rate, data.SlopeLower, data.SlopeUpper)
-			}
-
-			t.Logf("Rate %.1f → %s (slope=%.6e, CI=[%.6e, %.6e], peak=%d, mean=%.1f)",
-				tt.rate, report.Classification, data.Slope, data.SlopeLower, data.SlopeUpper,
-				data.PeakInFlight, data.MeanInFlight)
-		})
-	}
-}
-
-// generateTestWorkload creates a simplified workload for saturation testing.
-// Mimics the structure of requests generated by the simulator.
-func generateTestWorkload(rate float64, numRequests int, horizonUs int64) []*sim.Request {
-	requests := make([]*sim.Request, 0, numRequests)
-	iatUs := int64(1_000_000 / rate)
-
-	// Simple model: constant service time with FIFO queuing
-	serviceTimeUs := int64(100_000) // 100ms
-	serverFreeAt := int64(0)
-
-	for i := 0; i < numRequests; i++ {
-		arrivalUs := int64(i) * iatUs
-
-		// Stop if horizon specified and we've reached it
-		if horizonUs > 0 && arrivalUs >= horizonUs {
-			break
-		}
-
-		// FIFO queue
-		serviceStartUs := arrivalUs
-		if serverFreeAt > arrivalUs {
-			serviceStartUs = serverFreeAt
-		}
-
-		completionUs := serviceStartUs + serviceTimeUs
-		serverFreeAt = completionUs
-
-		// Determine state based on horizon
-		var state sim.RequestState
-		var ttftSet bool
-		var itl []int64
-
-		if horizonUs > 0 && completionUs > horizonUs {
-			// Doesn't complete within observation
-			state = sim.StateRunning
-			ttftSet = false
-			itl = nil
-		} else {
-			// Completes
-			state = sim.StateCompleted
-			ttftSet = true
-			itl = []int64{serviceTimeUs / 2}
-		}
-
-		requests = append(requests, &sim.Request{
-			ID:             fmt.Sprintf("req_%d", i),
-			ArrivalTime:    arrivalUs,
-			FirstTokenTime: serviceTimeUs / 2,
-			TTFTSet:        ttftSet,
-			ITL:            itl,
-			State:          state,
-			InputTokens:    make([]int, 100),
-			OutputTokens:   make([]int, 50),
-		})
-	}
-
-	return requests
-}
