+++ sim/workload/tracev2.go
@@ -26,7 +26,7 @@ type TraceHeader struct {
 	//   - a file path when --workload-spec is used (e.g. "workload.yaml")
 	//   - "preset:<name>" when --workload is used (e.g. "preset:chatbot")
 	//   - empty (omitted) when distribution synthesis or concurrency mode is used
-	WorkloadSpec string `yaml:"workload_spec,omitempty"`
+	WorkloadSpec   string `yaml:"workload_spec,omitempty"`
 
 	Server  *TraceServerConfig  `yaml:"server,omitempty"`
 	Network *TraceNetworkConfig `yaml:"network,omitempty"`
@@ -34,13 +34,13 @@ type TraceHeader struct {
 
 // TraceServerConfig captures server configuration in trace header.
 type TraceServerConfig struct {
-	Type                 string  `yaml:"type,omitempty"`
-	Model                string  `yaml:"model,omitempty"`
-	TensorParallel       int     `yaml:"tensor_parallel,omitempty"`
-	MaxNumSeqs           int     `yaml:"max_num_seqs,omitempty"`
-	BlockSize            int     `yaml:"block_size,omitempty"`
-	GPUMemoryUtilization float64 `yaml:"gpu_memory_utilization,omitempty"`
-	MaxModelLen          int64   `yaml:"max_model_len,omitempty"`
+	Type                  string  `yaml:"type,omitempty"`
+	Model                 string  `yaml:"model,omitempty"`
+	TensorParallel        int     `yaml:"tensor_parallel,omitempty"`
+	MaxNumSeqs            int     `yaml:"max_num_seqs,omitempty"`
+	BlockSize             int     `yaml:"block_size,omitempty"`
+	GPUMemoryUtilization  float64 `yaml:"gpu_memory_utilization,omitempty"`
+	MaxModelLen           int64   `yaml:"max_model_len,omitempty"`
 }
 
 // TraceNetworkConfig captures network configuration in trace header.
@@ -54,7 +54,7 @@ type TraceRecord struct {
 	ClientID          string
 	TenantID          string
 	SLOClass          string
-	VLLMPriority      int // vLLM priority value (0=highest urgency, higher=lower urgency); 0 when not set
+	VLLMPriority      int    // vLLM priority value (0=highest urgency, higher=lower urgency); 0 when not set
 	SessionID         string
 	RoundIndex        int
 	PrefixGroup       string
@@ -176,7 +176,7 @@ func ExportTraceV2(header *TraceHeader, records []TraceRecord, headerPath, dataP
 			r.Model,
 			strconv.FormatInt(r.DeadlineUs, 10),
 			strconv.Itoa(r.ServerInputTokens),
-			strconv.FormatInt(r.ArrivalTimeUs, 10),    // integer format
+			strconv.FormatInt(r.ArrivalTimeUs, 10),   // integer format
 			strconv.FormatInt(r.SendTimeUs, 10),       // integer format
 			strconv.FormatInt(r.FirstChunkTimeUs, 10), // integer format
 			strconv.FormatInt(r.LastChunkTimeUs, 10),  // integer format
@@ -488,7 +488,7 @@ func RequestsToTraceRecords(requests []*sim.Request) []TraceRecord {
 			PrefixGroup:      req.PrefixGroup,
 			PrefixLength:     prefixLen,
 			Streaming:        req.Streaming,
-			InputTokens:      inputTokens,           // suffix-only: total - PrefixLength
+			InputTokens:      inputTokens,      // suffix-only: total - PrefixLength
 			OutputTokens:     len(req.OutputTokens), // pre-determined count for replay fidelity
 			TextTokens:       req.TextTokenCount,
 			ImageTokens:      req.ImageTokenCount,
